Updates on reading from Bluetooth socket
InputStream.read() is a blocking call, it will wait until something has been put into the stream. So, there is no need to deliberately ask the thread to sleep. (I guess the sleeping was designed to wait for OBD to response?) 
In my test the average response time for a single command is about 67ms.
BTW, good work!
Yunfei